ENG-14774: Try to create optimal protection groups
The goal is to create the smallest protection groups as possible that
still honors ha groups. A bonus goal is that the protection group which
contains the partitions with the highest IDs is one of the smallest
groups.

Hosts are selected by a selector implementation which is chosen based
upon the ha group layout. Once a protection group is created it is
filled with the highest unused partitions.

The public API of AbstractTopology is simplified to 3 methods initial
topology creation, adding new hosts and a host rejoin.